+++ packages/fondue/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @frontify/fondue
 
+## 13.3.2
+
+### Patch Changes
+
+- [#2672](https://github.com/Frontify/fondue/pull/2672) [`b8744a7`](https://github.com/Frontify/fondue/commit/b8744a79da349bf9c34f25f826546bddc7678207) Thanks [@noahwaldner](https://github.com/noahwaldner)! - fix: align prop types in DatePicker
+
+- [#2668](https://github.com/Frontify/fondue/pull/2668) [`253d60e`](https://github.com/Frontify/fondue/commit/253d60e13fc9f34d089b199ceeb06dea42e47453) Thanks [@noahwaldner](https://github.com/noahwaldner)! - fix: fix text wrapping issue in datepicker input
+
+- Updated dependencies [[`b8744a7`](https://github.com/Frontify/fondue/commit/b8744a79da349bf9c34f25f826546bddc7678207), [`4e620ee`](https://github.com/Frontify/fondue/commit/4e620ee93572f1fc47859008a21a50bd96a53fe8), [`d9c4843`](https://github.com/Frontify/fondue/commit/d9c4843c9eae13e54a78654c8ce714f8d030c52d), [`253d60e`](https://github.com/Frontify/fondue/commit/253d60e13fc9f34d089b199ceeb06dea42e47453)]:
+    - @frontify/fondue-components@29.1.0
+    - @frontify/fondue-rte@0.1.2
+    - @frontify/fondue-tokens@5.0.0
+
+## 13.3.1
+
+### Patch Changes
+
+- [`4941eab`](https://github.com/Frontify/fondue/commit/4941eab80bbe9a588c946ab6e69bec3f7f9634e3) Thanks [@noahwaldner](https://github.com/noahwaldner)! - fix: minor datepicker fixes
+
+- Updated dependencies [[`d70403d`](https://github.com/Frontify/fondue/commit/d70403d07d7e4890666c4950c5f272b02c36953c), [`164a1fb`](https://github.com/Frontify/fondue/commit/164a1fb455d3d796fded75508fbd2eb203934dd1)]:
+    - @frontify/fondue-components@29.0.1
+    - @frontify/fondue-tokens@5.0.0
+
 ## 13.3.0
 
 ### Minor Changes